requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.340
transformers==4.35.2
torch==2.2.0
pydantic==2.5.2
sentence-transformers==2.2.2
chromadb==0.4.18
sqlalchemy==1.4.41
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp==1.29.0
stripe==7.11.0
jinja2==3.1.2
httpx==0.25.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
optree>=0.13.0
psycopg2-binary==2.9.6
